// src/models/contactModel.ts
export type Contact = {
  id?: number | string;
  name?: string;
  email?: string;
  apellido?: string;
  first_name?: string;
  last_name?: string;
  phone?: string;
  telefono?: string;
  phone_2?: string;
  mobile?: string;
  customer_id?: string;
  route_id?: string;
  address?: string;
  compound?: string;
  region?: string;
  visit_day?: string;
  recuring?: string;
  origin?: string;
  contact_source?: string;
  cedula_ruc?: string;
};