// src/db/models/dbContactModel.ts
export type DBContact = {
  id?: number;
  name?: string;
  email?: string;
  mobile?: string;
  phone?: string;
  city?: string;
  street?: string;
  street2?: string;
  x_telephone?: string;
  x_CustomerID?: string;
  x_Route?: string;
  x_compound?: string;
  x_CustomerStatus?: string;
  x_Day?: string;
};